+++ src/utils/formatFollowers.ts
@@ -1,6 +1,6 @@
 export function formatFollowers(num: number) {
-  return new Intl.NumberFormat('en', {
-    notation: 'compact',
+  return new Intl.NumberFormat("en", {
+    notation: "compact",
     maximumFractionDigits: 1
   }).format(num);
 }
